Move Work Item DTO's related fundtions to DTO's file
