+++ src/showWorkItems/WorkItemAndItsChildren.tsx
@@ -4,13 +4,13 @@ import React from "react"
 
 import { DisplayWorkItem } from "./DisplayWorkItem"
 import { useFetchWorkItem } from "./hooks"
-import { WorkItemDto, extractWorkItemId } from "./WorkItemDto"
+import { getChildrenIds, getParentId } from "./WorkItemDto"
 import { useWorkItemIdValue } from "@/contexts/WorkItemIdContext"
 
 
-export const WorkItemAndItsChildren: React.FC<{
-}> = ({
-}) => {
+export const WorkItemAndItsChildren: React.FC =
+  (
+  ) => {
     const workItemId = useWorkItemIdValue()
     if (workItemId == null) return
     return WorkItemAndItsChildrenInternal({ workItemId })
@@ -64,26 +64,3 @@ const WorkItemAndItsChildrenInternal: React.FC<{
       </div>
     )
   }
-
-function getParentId(workItemDto?: WorkItemDto): number | null {
-  const parentUrl = workItemDto
-    ?.relations
-    ?.find(x => x.rel === "System.LinkTypes.Hierarchy-Reverse")
-    ?.url
-
-  if (!parentUrl) return null
-  return extractWorkItemId(parentUrl)
-}
-
-function getChildrenIds(workItemDto?: WorkItemDto): number[] {
-  const maybeIds: (number | null)[] | undefined =
-    workItemDto
-      ?.relations
-      ?.filter(x => x.rel === "System.LinkTypes.Hierarchy-Forward")
-      .map(x => extractWorkItemId(x.url))
-  return omitFalsyValues(maybeIds ?? [])
-}
-
-function omitFalsyValues<T>(arg: (T | null | undefined)[]): T[] {
-  return arg.filter(x => !!x) as T[]
-}
