+++ src/showWorkItems/WorkItemDto.ts
@@ -29,3 +29,26 @@ export function extractWorkItemId(url?: string): number | null {
     const idChunk = url?.substring(index + WORK_ITEMS_URL.length + 1)
     return Number(idChunk)
 }
+
+export function getParentId(workItemDto?: WorkItemDto): number | null {
+    const parentUrl = workItemDto
+        ?.relations
+        ?.find(x => x.rel === "System.LinkTypes.Hierarchy-Reverse")
+        ?.url
+
+    if (!parentUrl) return null
+    return extractWorkItemId(parentUrl)
+}
+
+export function getChildrenIds(workItemDto?: WorkItemDto): number[] {
+    const maybeIds: (number | null)[] | undefined =
+        workItemDto
+            ?.relations
+            ?.filter(x => x.rel === "System.LinkTypes.Hierarchy-Forward")
+            .map(x => extractWorkItemId(x.url))
+    return omitFalsyValues(maybeIds ?? [])
+}
+
+function omitFalsyValues<T>(arg: (T | null | undefined)[]): T[] {
+    return arg.filter(x => !!x) as T[]
+}
